change IDENTITY columns from INT to BIGINT - required by Fabric DW

+++ elt-framework/ControlDB/ELT/Tables/ColumnMapping.sql
@@ -1,7 +1,7 @@
 ﻿GO
 
 CREATE TABLE [ELT].[ColumnMapping](
-	[MappingID] int NOT NULL IDENTITY,
+	[MappingID] bigint NOT NULL IDENTITY,
 	[IngestID] int NULL,
 	[L1TransformID] int NULL,
 	[L2TransformID] int NULL,
